requirements.txt
polars>=0.20.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
optuna>=3.5.0
joblib>=1.3.0
pyyaml>=6.0
streamlit>=1.32.0
plotly>=5.18.0
kagglehub>=0.2.0